ATR-662: refactoring - moved walker to a separate file
