+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/PXGraphCreateInstance/PXGraphCreateInstanceAnalyzer.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Immutable;
 
 using Acuminator.Utilities;
-using Acuminator.Utilities.DiagnosticSuppression;
 using Acuminator.Utilities.Roslyn.Semantic;
 
 using Microsoft.CodeAnalysis;
@@ -12,70 +11,8 @@ using Microsoft.CodeAnalysis.Diagnostics;
 namespace Acuminator.Analyzers.StaticAnalysis.PXGraphCreateInstance
 {
 	[DiagnosticAnalyzer(LanguageNames.CSharp)]
-	public class PXGraphCreateInstanceAnalyzer : PXDiagnosticAnalyzer
+	public partial class PXGraphCreateInstanceAnalyzer : PXDiagnosticAnalyzer
 	{
-		private class Walker : CSharpSyntaxWalker
-		{
-			private readonly SyntaxNodeAnalysisContext _context;
-			private readonly PXContext _pxContext;
-			private readonly SemanticModel _semanticModel;
-
-			private static readonly DiagnosticDescriptor _px1001Descriptor = Descriptors.PX1001_PXGraphCreateInstance;
-			private static readonly DiagnosticDescriptor _px1003Descriptor = Descriptors.PX1003_NonSpecificPXGraphCreateInstance;
-
-			public Walker(SyntaxNodeAnalysisContext context, PXContext pxContext, SemanticModel semanticModel)
-			{
-				_context = context;
-				_pxContext = pxContext;
-				_semanticModel = semanticModel;
-			}
-
-			public override void VisitObjectCreationExpression(ObjectCreationExpressionSyntax node)
-			{
-				_context.CancellationToken.ThrowIfCancellationRequested();
-
-				if (node.Type == null || _semanticModel.GetSymbolOrFirstCandidate(node.Type, _context.CancellationToken) is not ITypeSymbol typeSymbol)
-				{
-					base.VisitObjectCreationExpression(node);
-					return;
-				}
-
-				DiagnosticDescriptor? descriptor = GetDiagnosticDescriptor(typeSymbol);
-
-				if (descriptor != null)
-				{
-					_context.ReportDiagnosticWithSuppressionCheck(Diagnostic.Create(descriptor, node.GetLocation()),
-						_pxContext.CodeAnalysisSettings);
-				}
-
-				base.VisitObjectCreationExpression(node);
-			}
-
-			private DiagnosticDescriptor? GetDiagnosticDescriptor(ITypeSymbol typeSymbol)
-			{
-				if (typeSymbol is ITypeParameterSymbol typeParameterSymbol && typeParameterSymbol.IsPXGraph(_pxContext))
-				{
-					return _px1001Descriptor;
-				}
-				else if (typeSymbol.InheritsFrom(_pxContext.PXGraph.Type))
-				{
-					return _px1001Descriptor;
-				}
-				else if (typeSymbol.Equals(_pxContext.PXGraph.Type, SymbolEqualityComparer.Default))
-				{
-					return _px1003Descriptor;
-				}
-
-				return null;
-			}
-
-			public override void DefaultVisit(SyntaxNode node)
-			{
-				_context.CancellationToken.ThrowIfCancellationRequested();
-				base.DefaultVisit(node);
-			}
-		}
-
 		public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(
 				Descriptors.PX1001_PXGraphCreateInstance,
 				Descriptors.PX1003_NonSpecificPXGraphCreateInstance);
